+++ src/ui/GalleryLayer.hpp
@@ -65,6 +65,7 @@ public:
 	void onNavButton(CCObject *sender);
 	void onWebsite(CCObject *sender);
 	void onDiscord(CCObject *sender);
+	void onReset(CCObject *sender);
 
 	void refreshGallery();
 	void setupIconPack();